+++ API/Extensions/DependencyInjectionExtensions.cs
@@ -85,6 +85,9 @@ namespace API.Extensions
             services.AddScoped<IUserProjectLikeService, UserProjectLikeService>();
             services.AddScoped<IUserProjectLikeRepository, UserProjectLikeRepository>();
 
+            services.AddScoped<ICallToActionOptionService, CallToActionOptionService>();
+            services.AddScoped<ICallToActionOptionRepository, CallToActionOptionRepository>();
+
             return services;
         }
     }