+++ API/Extensions/UsersExtensions.cs
@@ -16,6 +16,7 @@
 */
 
 using API.Configuration;
+using API.InternalResources;
 using IdentityModel.Client;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -32,6 +33,7 @@ using System.Net.Http;
 using System.Security.Claims;
 using System.Security.Principal;
 using System.Threading.Tasks;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace API.Extensions
 {
@@ -104,9 +106,21 @@ namespace API.Extensions
         /// <param name="actionContext">The action context.</param>
         /// <param name="config">The configuration.</param>
         /// <returns>The user object with information retrieved from the identity server</returns>
-        public static User GetUserInformation(this HttpContext actionContext, Config config)
+        public static UserCreateInternalResource GetUserInformation(this HttpContext actionContext, Config config)
         {
             string bearerToken = actionContext.Request.Headers.GetCommaSeparatedValues("Authorization").FirstOrDefault();
+            string providerId = "";
+
+            if(bearerToken != null)
+            {
+                string token = bearerToken.Replace("Bearer ", "");
+                JwtSecurityTokenHandler handler = new JwtSecurityTokenHandler();
+                if(handler.ReadToken(token) is JwtSecurityToken tokens)
+                {
+                    providerId = tokens.Claims.FirstOrDefault(claim => claim.Type == "idp")?.Value;
+                }
+            }
+
             if(string.IsNullOrEmpty(bearerToken))
             {
                 return null;
@@ -123,11 +137,13 @@ namespace API.Extensions
             {
                 return null;
             }
-            User newUser = new User()
+
+            UserCreateInternalResource newUser = new UserCreateInternalResource
             {
                 Name = (string) jsonResponse["name"],
                 Email = (string) jsonResponse["email"],
-                IdentityId = (string) jsonResponse["sub"]
+                IdentityId = (string) jsonResponse["sub"],
+                IdentityInstitutionId = providerId
             };
             return newUser ;
         }