+++ Data/ApplicationDbContext.cs
@@ -15,9 +15,7 @@
 * If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
 */
 
-using System.Collections.Generic;
 using Data.Configurations;
-using Data.Helpers;
 using Microsoft.EntityFrameworkCore;
 using Models;
 
@@ -39,14 +37,31 @@ namespace Data
         /// The user.
         /// </value>
         public DbSet<User> User { get; set; }
-
+        
+        /// <summary>
+        /// Gets or sets the Call to Action.
+        /// </summary>
+        /// <value>
+        /// The call to action.
+        /// </value>
+        public DbSet<CallToAction> CallToAction { get; set; }
+        
         /// <summary>
         /// Gets or sets the project.
         /// </summary>
         /// <value>
         /// The project.
         /// </value>
+        public DbSet<File> File { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Call To Action options.
+        /// </summary>
+        /// <value>
+        /// The call to action options.
+        /// </value>
         public DbSet<Project> Project { get; set; }
+        
         /// <summary>
         /// Gets or sets the collaborators.
         /// </summary>
@@ -54,6 +69,7 @@ namespace Data
         /// The collaborators.
         /// </value>
         public DbSet<Collaborator> Collaborators { get; set; }
+        
         /// <summary>
         /// Gets or sets the highlight.
         /// </summary>
@@ -61,6 +77,7 @@ namespace Data
         /// The highlight.
         /// </value>
         public DbSet<Highlight> Highlight { get; set; }
+        
         /// <summary>
         /// Gets or sets the embedded project.
         /// </summary>
@@ -68,6 +85,7 @@ namespace Data
         /// The embedded project.
         /// </value>
         public DbSet<EmbeddedProject> EmbeddedProject { get; set; }
+        
         /// <summary>
         /// Gets or sets the role.
         /// </summary>
@@ -75,6 +93,7 @@ namespace Data
         /// The role.
         /// </value>
         public DbSet<Role> Role { get; set; }
+        
         /// <summary>
         /// Gets or sets the institution.
         /// </summary>
@@ -82,25 +101,50 @@ namespace Data
         /// The institution.
         /// </value>
         public DbSet<Institution> Institution { get; set; }
+        
+        /// <summary>
         /// Gets or sets the file.
         /// </summary>
         /// <value>
         /// The file.
         /// </value>
         public DbSet<File> File { get; set; }
+        
         /// <summary>
         /// Gets or sets the projects liked by users.
         /// </summary>
+        /// <value>
+        /// The like by the user.
+        /// </value>
         public DbSet<ProjectLike> ProjectLike { get; set; }
 
+        /// <summary>
+        /// Gets or sets the call to action option.
+        /// </summary>
+        /// <value>
+        /// The call to action option.
+        /// </value>
+        public DbSet<CallToActionOption> CallToActionOption { get; set; }
+
+        /// <summary>
+        /// Gets or sets the user follwoing the project.
+        /// </summary>
+        /// <value>
+        /// The user following the project.
+        /// </value>
         public DbSet<UserProject> UserProject { get; set; }
+
+        /// <summary>
+        /// Gets or sets user following the user.
+        /// </summary>
+        /// <value>
+        /// The user following the user.
+        /// </value>
         public DbSet<UserUser> UserUser { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             base.OnModelCreating(modelBuilder);
-
-            //modelBuilder.Properties<DateTime>().Configure(c => c.HasColumnType("datetime2"));
             modelBuilder.ApplyConfiguration(new UserConfiguration());
             modelBuilder.ApplyConfiguration(new ProjectConfiguration());
         }